+++ changelog.md
@@ -1,4 +1,6 @@
 # v0.1.0
+1. typing checking
+2. support routing path / pathprefix
 
 # v0.0.1
 1. 支持简单的 http 请求